Navbar is not a LiveComponent anymore
The navbar does not have any independent status so it doesn't need to
be a LiveComponent. Being a LiveComponent was actually problematic as
it was not able to generate the proper dashboard path for linking.
