+++ lib/error_tracker/web/components/layouts/live.html.heex
@@ -1,4 +1,4 @@
-<.live_component module={Navbar} id="navbar" {assigns} />
+<.navbar id="navbar" {assigns} />
 <main class="container px-4 mx-auto mt-4 mb-4">
   <%= @inner_content %>
 </main>
